+++ noodles-bam/src/reader/record.rs
@@ -1,6 +1,7 @@
 //! BAM record field readers.
 
 pub mod data;
+mod sequence;
 
 use std::{
     io::{self, Read},
@@ -38,9 +39,8 @@ where
 }
 
 pub(crate) fn decode_record(src: &mut BytesMut, record: &mut Record) -> io::Result<()> {
-    use super::alignment_record::{
-        get_cigar, get_data, get_quality_scores, get_read_name, get_sequence,
-    };
+    use self::sequence::get_sequence;
+    use super::alignment_record::{get_cigar, get_data, get_quality_scores, get_read_name};
 
     *record.reference_sequence_id_mut() = get_reference_sequence_id(src)?;
     *record.position_mut() = get_position(src)?;